Lots of changes, still need to allow user to update preferences
