+++ app/controllers/google_controller.rb
@@ -1,0 +1,208 @@
+require 'google/apis/sheets_v4'
+require 'googleauth'
+require 'googleauth/stores/file_token_store'
+
+require 'fileutils'
+
+class GoogleController < ApplicationController
+  include MatchesHelper
+  include ApplicationHelper
+  include GoogleHelper
+
+  before_filter :check_logged_in
+
+  def projects_groups_fetch
+    unless authorize(false).nil?
+      if fetch_project_data
+        fetch_group_data
+      end
+    end
+  end
+
+  def write_matches
+    write_all_matches
+  end
+
+  def fetch_matches
+    update_matches_sheet
+  end
+
+  def save_preferences
+    update_settings
+    redirect_to google_fetch_path
+  end
+
+  def index
+  end
+
+  def settings
+  end
+
+  #### CS169-PGM GOOGLE METHODS BEGIN ####
+
+  def fetch_project_data
+    response = fetch_data('Projects!A1:B50')
+    unless response.nil?
+      adjust_projects response
+      return true
+    end
+    false
+  end
+
+  def adjust_projects(response)
+    Project.destroy_all
+    delete_matches(response)
+    create_projects(response)
+  end
+
+  def create_projects(response)
+    response.values.each do |row|
+      Project.create(:id => row[0], :project_name => row[1])
+    end
+  end
+
+  #Fetches the data from the google sheet
+  def fetch_group_data
+    response = fetch_data('Groups!A1:K29')
+    unless response.nil?
+      adjust_groups response
+    end
+  end
+
+  # Takes the google sheet response and generates all the groups from it
+  def adjust_groups(response)
+    Group.destroy_all
+    delete_matches(response)
+    create_groups(response)
+    redirect_to root_path
+  end
+
+  def create_groups(response)
+    response.values.each do |row|
+      preferences = []
+      preference_rows = row.drop(4)
+      preference_rows.each do |preference|
+        preferences.push((/\d+/.match(preference))[0])
+      end
+      Group.create(:id => row[2], :created_at => row[0], :username => row[1],
+                   :first_choice => preferences[0], :second_choice => preferences[1], :third_choice => preferences[2],
+                   :fourth_choice => preferences[3], :fifth_choice => preferences[4], :sixth_choice => preferences[5],
+                   :seventh_choice => preferences[6])
+    end
+  end
+
+  def update_matches_sheet
+    response = fetch_data('Current_Match!A1:C50')
+    unless response.nil?
+      response = delete_matches(response)
+      response.values.each do |row|
+        if row[0] != 'None'
+          Match.create(:group_id => row[0], :project_id => row[1])
+        end
+      end
+      flash[:notice] = 'Matches Fetched and '
+      write_all_matches
+    end
+  end
+
+  def write_all_matches
+    temp_matches = Match.all.sort_by { |some_match| (some_match.group.id.to_i) }
+    match_values = [['Group ID', 'Project ID', 'Project Name']]
+    temp_matches.each do |match|
+      match_values.push([match.group.id, match.project.id, match.project.project_name])
+    end
+    find_unmatched_projects.each do |project|
+      match_values.push(['None', project.id, project.project_name])
+    end
+    range = 'Current_Match!A1:C50'
+    begin
+      service_authorize.clear_values(current_parameters[:spreadsheet_id], range)
+      values = Google::Apis::SheetsV4::ValueRange.new({:values => match_values, :major_dimension => 'ROWS'})
+      service_authorize.update_spreadsheet_value(current_parameters[:spreadsheet_id], range, values, value_input_option:'USER_ENTERED')
+      if flash[:notice]
+        flash[:notice] += 'Matches Re-Alphabetized and Written.'
+      else
+        flash[:notice] = 'Matches Written.'
+      end
+      redirect_to google_fetch_path
+    rescue Signet::AuthorizationError
+      authorize true
+      return
+    rescue Google::Apis::AuthorizationError
+      authorize true
+      return
+    end
+  end
+
+  #### GOOGLE METHODS BEGIN ####
+
+  OOB_URI = 'urn:ietf:wg:oauth:2.0:oob'
+  APPLICATION_NAME = 'CS169'
+  CLIENT_SECRETS_PATH = 'client_secret.json'
+  CREDENTIALS_PATH = File.join(Dir.home, '.credentials',
+                               "sheets.googleapis.com-ruby-quickstart.yaml")
+  SCOPE = Google::Apis::SheetsV4::AUTH_SPREADSHEETS
+
+  def complete_authorization
+    FileUtils.mkdir_p(File.dirname(CREDENTIALS_PATH))
+
+    client_id = Google::Auth::ClientId.from_file(CLIENT_SECRETS_PATH)
+    token_store = Google::Auth::Stores::FileTokenStore.new(file: CREDENTIALS_PATH)
+    authorizer = Google::Auth::UserAuthorizer.new(
+        client_id, SCOPE, token_store)
+    user_id = 'default'
+
+    begin
+      authorizer.get_and_store_credentials_from_code(user_id: user_id, code: params[:code], base_url: OOB_URI)
+      session[:is_authorized] = true
+    rescue Signet::AuthorizationError
+      flash[:notice] = 'Error, invalid code. Try again'
+    end
+
+    redirect_to google_fetch_path
+  end
+
+  ##
+  # Ensure valid credentials, either by restoring from the saved credentials
+  # files or intitiating an OAuth2 authorization. If authorization is required,
+  # the user's default browser will be launched to approve the request.
+  #
+  # @return [Google::Auth::UserRefreshCredentials] OAuth2 credentials
+  def authorize(force_reload)
+    FileUtils.mkdir_p(File.dirname(CREDENTIALS_PATH))
+
+    client_id = Google::Auth::ClientId.from_file(CLIENT_SECRETS_PATH)
+    token_store = Google::Auth::Stores::FileTokenStore.new(file: CREDENTIALS_PATH)
+    authorizer = Google::Auth::UserAuthorizer.new(
+        client_id, SCOPE, token_store)
+    user_id = 'default'
+    credentials = authorizer.get_credentials(user_id)
+    if force_reload || credentials.nil?
+      session[:is_authorized] = false
+      redirect_to google_fetch_path
+      return
+    end
+    credentials
+  end
+
+  def service_authorize
+    service = Google::Apis::SheetsV4::SheetsService.new
+    service.client_options.application_name = APPLICATION_NAME
+    service.authorization = authorize false
+    service
+  end
+
+  def fetch_data(range)
+    begin
+      response = service_authorize.get_spreadsheet_values(current_parameters[:spreadsheet_id], range)
+    rescue Signet::AuthorizationError
+      authorize true
+      return
+    rescue Google::Apis::AuthorizationError
+      authorize true
+      return
+    end
+    response
+  end
+
+end
