--- app/views/google/complete_authorization.html.erb
@@ -1,9 +1,0 @@
-<%= render "layouts/sidebar.html.erb" %>
-
-<div class="content-container">
-  <div class="content">
-    <div class="action-container">
-      COMPLETE AUTHORIZATION
-    </div>
-  </div>
-</div>
