--- app/views/google/get_authorization.html.erb
@@ -1,18 +1,0 @@
-<%= render "layouts/sidebar.html.erb" %>
-
-<div class="content-container">
-  <div class="content">
-    <div class="action-container">
-      <% if flash[:notice] %>
-          <%= flash.notice %>
-      <% end %>
-      <br>
-      Go to the url and copy and paste the code into this box:
-      <a href="<%= session[:authorize] %>">Google Authorization</a>
-        <%= form_tag(google_authorize_path, method: "POST") do %>
-          Code: <input type="text" name="code" id="code"><br>
-          <input id="submit" class="button" type="submit" value="Submit Code"/>
-      <% end %>
-    </div>
-  </div>
-</div>
