--- app/views/google/show.html.erb
@@ -1,9 +1,0 @@
-<%= render "layouts/sidebar.html.erb" %>
-
-<div class="content-container">
-  <div class="content">
-    <div class="action-container">
-      Hello World
-    </div>
-  </div>
-</div>
