+++ app/views/layouts/_sidebar.html.erb
@@ -15,7 +15,7 @@
                     <a class="nav-link" href="<%= projects_path()%>">Projects</a>
                 </li>
                 <li class="nav-li">
-                  <a class="nav-link" href="<%= google_fetch_path %>">Fetch Data</a>
+                  <a class="nav-link" href="<%= google_fetch_path %>">Google Sheets</a>
                 </li>
             <% end %>
         </ul>
