+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20161125215839) do
+ActiveRecord::Schema.define(version: 20161129005532) do
 
   create_table "groups", force: :cascade do |t|
     t.string   "group_name"
